solved conflicts and warnings
