+++ frontend/better-you/src/assets/scss/dashboard-page/GeneralGoalPopupStyle.scss
@@ -1,4 +1,4 @@
-@import "./_colors.scss";
+@import "../globals/colors";
 
 .genera-popup{
     background-color: rgba(0,0,0,0.6);
