+++ frontend/better-you/src/assets/scss/dashboard-page/GeneralGoalViewStyle.scss
@@ -1,4 +1,4 @@
-@import "./_colors.scss";
+@import "../globals/colors";
 
 .general-goal-container{
     width: 55vw;
