+++ frontend/better-you/src/assets/scss/settings/ResetPasswordStyle.scss
