+++ frontend/better-you/src/components/dashboard-page/lists/habits/HabitCard.tsx
@@ -6,7 +6,7 @@ import Typography from "@material-ui/core/Typography";
 import Tooltip from "@material-ui/core/Tooltip";
 import DeleteIcon from "@material-ui/icons/Delete";
 import IconButton from "@material-ui/core/IconButton";
-import "../assets/scss/GoalListStyle.scss";
+import "../../../../assets/scss/dashboard-page/GoalListStyle.scss";
 
 interface IProps {
   habit: { title: string; repetitionType: string };
