+++ frontend/better-you/src/components/dashboard-page/lists/newsfeed/NewsfeedCard.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Card from "@material-ui/core/Card";
-import "../assets/scss/NewsfeedListStyle.scss";
+import "../../../../assets/scss/dashboard-page/NewsfeedListStyle.scss";
 import Fab from "@material-ui/core/Fab";
 import AccountCircleOutlinedIcon from "@material-ui/icons/AccountCircleOutlined";
 import Tooltip from "@material-ui/core/Tooltip";
@@ -14,16 +14,12 @@ interface IProps {
 }
 
 class NewsfeedCard extends React.Component<IProps, {}> {
-  constructor(prop: IProps) {
-    super(prop);
-  }
-
   render() {
     let stil = {
       backgroundImage: 'url("' + this.props.image + '")'
     };
 
-    if (this.props.image == "")
+    if (this.props.image === "")
       return (
         <Card className="newsfeed_container">
           <Tooltip title="Go to profile">
